+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add `IN` operator
 * Add `NOT IN` operator
 * Remove `CONTAINS` operator
+* Add support for function calls
 
 ## 1.2.0 (2019-04-11)